+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-11592 Fixed missing translation of an empty cluster selector in Onepanel REST token template tile.
 * VFS-11575 Increased max zoom of jvectormap
 * VFS-11251 Removed unused oneicons
 * VFS-11134 Added information about membership to leaving modal